refactor: 各featureで使用するコンポーネントをまとめたcomponentパッケージを作成し、そこで保存するように変更
