+++ composeApp/src/commonMain/kotlin/org/example/project/ui/feature/profile/component/ToolsSection.kt
@@ -1,4 +1,4 @@
-package org.example.project.ui.feature.profile
+package org.example.project.ui.feature.profile.component
 
 import androidx.compose.foundation.Image
 import androidx.compose.foundation.layout.Arrangement
@@ -11,6 +11,8 @@ import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import org.example.project.model.Tool
 import org.example.project.model.ToolSet
+import org.example.project.ui.feature.profile.SectionContent
+import org.example.project.ui.feature.profile.SectionTitle
 import org.example.project.ui.theme.dimensions.IconSizes
 import org.example.project.ui.theme.dimensions.Paddings
 import org.jetbrains.compose.resources.painterResource
